@@ -1,20 +1,20 @@
 name: liqpay
 description: LiqPay payment service wrapper library to be used on a client or server side dart.
-version: 0.0.2
+version: 0.0.3
 homepage: https://github.com/Newgarden-Solutions/liqpay
 repository: https://github.com/Newgarden-Solutions/liqpay
 issue_tracker: https://github.com/Newgarden-Solutions/liqpay/issues
 
 environment:
-  sdk: ">=3.4.0 <4.0.0"
+  sdk: ">=3.9.0 <4.0.0"
 
 dependencies:
-  crypto: ^3.0.5
-  http: ^1.2.2
+  crypto: ^3.0.6
+  http: ^1.5.0
   http_interceptor: ^2.0.0
 
 dev_dependencies:
-  lints: ^4.0.0
-  test: ^1.25.8
-  mockito: ^5.4.4
-  build_runner: ^2.4.12
+  lints: ^6.0.0
+  test: ^1.26.3
+  mockito: ^5.5.1
+  build_runner: ^2.9.0
